+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - Changed to recommended JEI 1.16.5:7.7.1.153
 - Changed max unque bullet types to 18
 - Fixed Diamond and Armor Piercing bullet recipes showing 4 for craft to 8
+- Added Gunpowder recipe